+++ imio/dms/mail/testing.py
@@ -33,6 +33,7 @@ from Products.CMFPlone.utils import _createObjectByType
 from Products.CMFPlone.utils import base_hasattr
 from Products.CMFPlone.utils import safe_unicode
 from Products.ExternalMethod.ExternalMethod import manage_addExternalMethod
+from Products.GenericSetup.utils import PropertyManagerHelpers
 from profilehooks import timecall
 from Testing import ZopeTestCase as ztc
 # from z3c.relationfield import RelationValue
@@ -186,6 +187,9 @@ class DmsmailLayer(PloneWithPackageLayer):
         setRoles(portal, TEST_USER_ID, ["Member"])
 
     def setUpZope(self, app, configurationContext):
+        # zope.conf's `default-zpublisher-encoding utf-8` isn't read in tests, so GenericSetup
+        # re-encodes xml profile values in iso-8859-15 (Bad Parapheo action title)
+        PropertyManagerHelpers._encoding = "utf-8"
         ztc.utils.setupCoreSessions(app)
         super(DmsmailLayer, self).setUpZope(app, configurationContext)
         from App.config import _config
@@ -226,6 +230,12 @@ class DmsmailLayerNP1(DmsmailLayer):
         )
         portal.portal_setup.runImportStepFromProfile(
             "profile-imio.dms.mail:singles", "imiodmsmail-task_n_plus_1_wfadaptation", run_dependencies=False
+        )
+        portal.portal_setup.runImportStepFromProfile(
+            "profile-imio.dms.mail:singles", "imiodmsmail-activate-om-signing", run_dependencies=False
+        )
+        portal.portal_setup.runImportStepFromProfile(
+            "profile-imio.dms.mail:singles", "imiodmsmail-activate-sign-request", run_dependencies=False
         )
         # Delete om
         brains = api.content.find(portal_type="dmsoutgoingmail")